Refactor teleop
