--- command/cpp/build.sh
@@ -1,6 +1,0 @@
-mkdir -p build
-cd build
-cmake ..
-make
-cp command_parser.x ../
-cp serial_reader.x ../
