+++ requirements.txt
@@ -3,6 +3,7 @@ numpy==1.13.3
 pygame==1.9.3
 pyserial==3.4
 pyzmq==16.0.2
+robust-serial==0.1
 scikit-learn==0.19.0
 scipy==0.19.1
 six==1.11.0
