+++ teleop/teleop_server.py
@@ -10,12 +10,14 @@ except ImportError:
     import Queue as queue
 
 import zmq
-import serial
-import picamera
+try:
+    import picamera
+except ImportError:
+    raise ImportError("Picamera package not found, you must run this code on the Raspberry Pi")
 
-import common
-from common import is_connected, CommandThread, ListenerThread, BAUDRATE, \
-    Order, sendOrder, n_received_semaphore, get_serial_ports, command_queue
+from robust_serial import write_order, Order
+from robust_serial.threads import CommandThread, ListenerThread
+from robust_serial.utils import open_serial_port, CustomQueue
 
 emptyException = queue.Empty
 fullException = queue.Full
@@ -35,15 +37,15 @@ if record_video:
     print("Recording a video to {}".format(args.video_file))
 
 try:
-    serial_port = get_serial_ports()[0]
-    serial_file = serial.Serial(port=serial_port, baudrate=BAUDRATE, timeout=0, writeTimeout=0)
+    serial_file = open_serial_port(baudrate=115200)
 except Exception as e:
     raise e
 
-# Wait until we are connected to the arduino
+is_connected = False
+# Initialize communication with Arduino
 while not is_connected:
     print("Waiting for arduino...")
-    sendOrder(serial_file, Order.HELLO.value)
+    write_order(serial_file, Order.HELLO)
     bytes_array = bytearray(serial_file.read(1))
     if not bytes_array:
         time.sleep(2)
@@ -52,14 +54,27 @@ while not is_connected:
     if byte in [Order.HELLO.value, Order.ALREADY_CONNECTED.value]:
         is_connected = True
 
+print("Connected to Arduino")
+
+# Create Command queue for sending orders
+command_queue = CustomQueue(2)
+# Number of messages we can send to the Arduino without receiving an acknowledgment
+n_messages_allowed = 3
+n_received_semaphore = threading.Semaphore(n_messages_allowed)
+# Lock for accessing serial file (to avoid reading and writing at the same time)
+serial_lock = threading.Lock()
+
+# Event to notify threads that they should terminate
 exit_event = threading.Event()
 
-threads = [CommandThread(serial_file, command_queue, exit_event),
-           ListenerThread(serial_file, exit_event)]
+print("Starting Communication Threads")
+# Threads for arduino communication
+threads = [CommandThread(serial_file, command_queue, exit_event, n_received_semaphore, serial_lock),
+           ListenerThread(serial_file, exit_event, n_received_semaphore, serial_lock)]
 for t in threads:
     t.start()
 
-print("Connected to Arduino, waiting for client...")
+print("Waiting for client...")
 socket.send(b'1')
 print("Connected To Client")
 i = 0
@@ -73,8 +88,8 @@ with picamera.PiCamera() as camera:
         control_speed, angle_order = socket.recv_json()
         print("({}, {})".format(control_speed, angle_order))
         try:
-            common.command_queue.put_nowait((Order.MOTOR, control_speed))
-            common.command_queue.put_nowait((Order.SERVO, angle_order))
+            command_queue.put_nowait((Order.MOTOR, control_speed))
+            command_queue.put_nowait((Order.SERVO, angle_order))
         except fullException:
             print("Queue full")
 
@@ -85,16 +100,16 @@ with picamera.PiCamera() as camera:
         camera.stop_recording()
 
 print("Sending STOP order...")
-# SEND STOP ORDER at the end
-common.resetCommandQueue()
+# Stop the car at the end
+command_queue.clear()
 n_received_semaphore.release()
-common.command_queue.put((Order.MOTOR, 0))
+command_queue.put((Order.MOTOR, 0))
 # Make sure STOP order is sent
 time.sleep(0.2)
 # End the threads
 exit_event.set()
 n_received_semaphore.release()
-print("EXIT")
+print("Exiting...")
 
 for t in threads:
     t.join()
